chore: Move all remaining actions into links (#37829)
The chain+link framework is now used for all the steps in the backup
command.


## Automation

/test sanity

### :mag: Cypress test results
