+++ app/client/packages/rts/src/ctl/backup/links/ManifestLink.ts
@@ -1,9 +1,12 @@
-import type { Link } from "./index";
+import type { Link } from ".";
 import type { BackupState } from "../BackupState";
 import * as utils from "../../utils";
 import fsPromises from "fs/promises";
 import path from "path";
 
+/**
+ * Creates a manifest file that contains metadata about the backup.
+ */
 export class ManifestLink implements Link {
   constructor(private readonly state: BackupState) {}
